Build templates based on many warped renditions
New MATLAB function `build_template`, that depends on private functions from the find-audio library, to warp together multiple spectrograms and create a template that can be used for matching. Also returns a set of weights that can be used to discount frequenty areas of variability (not currently used).

Updated DTM and match syllables to support loading spectral template directly.

+++ BelaWarpDetect/main.cpp
@@ -16,7 +16,7 @@
 std::vector<std::vector<float>> result_score;
 std::vector<std::vector<int>> result_length;
 
-void cbPrintResult(int index, float score, int length) {
+void cbPrintResult(size_t index, float score, int length) {
     std::cout << "Syllabe " << index << ": " << score << " (" << length << ")\n";
 }
